fix some bugs
